refactor: fix nwd structure
